Fixed Sharedpreferences return garbage after few runs
Sharedpreferences return garbage after few runs. It is also mentioned in https://developer.android.com/reference/android/content/SharedPreferences
"Caution: The preference manager does not currently store a strong reference to the listener. You must store a strong reference to the listener, or it will be susceptible to garbage collection. We recommend you keep a reference to the listener in the instance data of an object that will exist as long as you need the listener."